requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.3.0
langchain>=0.1.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
jupyter>=1.0.0
black>=23.0.0
flake8>=6.0.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0.0